use protocols instead of ABC Fixes #55
