+++ yogaflow/writer/abstract_writer.py
@@ -1,16 +1,23 @@
 """ Abstract output generator module """
 from abc import ABC, abstractmethod
 import os
+from typing import Protocol
 from yogaflow.yoga.yoga_class import YogaClass
 from yogaflow.config import get as get_config
 
-
-class AbstractWriter(ABC): #pylint: disable=R0903
+class Writer(Protocol):
     """ Abstract file writer class
     In case you need to generate a different output in the future;
     you can simply derive a new class and use in your project.
     """
+    def write(self, generated_class: YogaClass):
+        """ Generates the desired output """
 
+class AbstractWriter(ABC, Writer): #pylint: disable=R0903
+    """ Abstract file writer class
+    In case you need to generate a different output in the future;
+    you can simply derive a new class and use in your project.
+    """
     @abstractmethod
     def write(self, generated_class: YogaClass):
         """ Generates the desired output """
